refactor(tests): move fixtures to conftest.py under tests/bin/
Extracted hal_module and hal_instance fixtures from test_hal.py into
a conftest.py, relocating the test suite to tests/bin/ to better
reflect the file under test.
